merge fix for bug #4834 by gurer.
the right way to calculate installedsize is to iterate through self.files of the current package.
A source package may produce multiple packages as in skim-tables src package. Iterating through 
package.files (<Path> s of pspec.xml) and calculating the size with util.dir_size is not the 
right approach.

Some files that are in other produced packages may also be added to an unrelated package's - that 
actually has not that file - in its size calculation approach.