The current search method in package-manager takes 0.3 seconds to search letter "a" in all package database (name, summary, description) but to create the page and make it ready takes a total time of 3.8 seconds on an Intel Core Duo 1.66 Ghz machine.
Also tried an approach to improve search with caching, takes 0.03 sec
to find the results. But it took again ~3.6 seconds to prepare the
result page.

So it is not the search that is slow...

--------------

from kde/3.5/kdelibs/khtml/ecma/kjs_proxy.cpp:

  KJSCPUGuard guard; 
  guard.start(); 
  Completion comp =  m_script->evaluate(code, thisNode); 
  guard.stop();

...

guard is a timer. When the script to evaluate takes too much CPU time,
alarmHandler calls confirmTerminate:

bool KJSCPUGuard::confirmTerminate() 
{ 

  kdDebug(6070) << "alarmhandler"  << endl;
  return KMessageBox::warningYesNo(0L, i18n("A script on this page is causing KHTML to freeze. 
                                             If it continues to run, other applications may become 
					     less responsive.\nDo you want to abort the script?"), 
				       i18n("JavaScript"), i18n("&Abort"), KStdGuiItem::cont(),
				       "kjscupguard_alarmhandler") == KMessageBox::Yes;
}

----------------

The problem is in initShowHideDivs() of animation.js

it was been called at the end of the created html page.

         part.write(''' 
	 <script type="text/javascript">
         initShowHideDivs(); </script></body></html> 
	 ''')

It is a loop that adds attributes to all the DIV and INPUT
elements. But when the element arrays are relatively large, it hogs
precious CPU time. And if the CPU is a slow one and also busy working
with some other apps, it takes too much time to evaluate the js
script, thus guard calls terminate function.

We do not need to call initShowHideDivs, we are creating the html
page. So we add the attributes on the fly while creating the page.

Thanks to Firebug ;) also mdakin for introducing with it. :)

BUG:FIXED:5308