Do not keep piksemel node references in memory
I was worried about pisi-db's memory usage cause now we have everything in memory. Right?
Decided to look what was the memory usage of trunk/pisi? What do you expect from a database
implementation. Less memory usage, fast access?

We have below repositories added in two tests

pardus-2007.2
   http://paketler.pardus.org.tr/pardus-2007/pisi-index.xml.bz2
contrib-2007
   http://paketler.pardus.org.tr/contrib-2007/pisi-index.xml.bz2
contrib-2007-src
   http://svn.pardus.org.tr/contrib/pisi-index.xml.bz2
pardus-2007-src
   http://svn.pardus.org.tr/pardus/2007/pisi-index.xml.bz2

---------------------

trunk/pisi test

we run pisi la and check memory usage from "ps aux | grep pisi".
This is the average of memory when pisi la is running.

61300 51688 pts/4    R+   14:22   0:01 /usr/bin/python /usr/bin/pisi la

VMM: 61300 RSS: 51688

---------------------

branches/pisi-db

does not have a db, all the repositories are in _memory_ but still:
This is the average of memory when pisi la is running.

21752 18044 pts/4    R+   14:16   0:05 /usr/bin/python /usr/bin/pisi la

VMM: 21752 18044

--------------------

I see that the memory usage of branches/pisi that worried me was the same
with the trunk/pisi.

With this commit we reduce the usage by not keeping piksemel node references
by rev_dep dbs. But, I do not see the problem of trunk/pisi.