Compress further package - source xml in memory strings
Most of the memory is taken by package and source xml strings. We can
compress them in memory. Seems fast and harmless and not cause performance
problems.

below is the previous branches/pisi-db "pisi la" memory usage average:
21752 18044 pts/4    R+   14:16   0:05 /usr/bin/python /usr/bin/pisi la

below is with this commit:
15788 12124 pts/0    R+   14:46   0:07 /usr/bin/python /usr/bin/pisi la

This optimization may be removed if it causes any performance problems. But
so far so good.