Check for reinstalls as in install_pkg_names
This is actually added for YALI. YALI now gives all the packages with
their full paths. But when an exception happens and retry is selected
from YALI, to continue the packages where YALI left pisi has to check
the list and should not install the already installed packages as done
in install_pkg_names.