actionsapi/texlivemodules: Modified to work properly with TexLive 2009 version
Texlive 2009 has some significant changes in the source tree. 

-The format.cnf file is created from the information of tlpobj files. This is parsed via
the function addFormat().
-buildFormatFiles() has some enviroment changes.
-The files of the source itself is distributed in the source dir. This must be relocated 
and re-distributed properly. This relocation information is available in the the tlpobj files. 
The function which reads the information and relocate the files is named as  moveSource().
-handleConfigFiles() is rewritten to provide simplicity and readability. Also there is no need of 
parameters, because the prefixes are the same everywhere. 
-The order of the functions has been changed. First config files has to be created, after that it's 
possible to build formats