specfile: Replace "current" string in release/version values
With this feature, it is not necessary to hardcode versions/releases
used in inter-dependencies.

    <RuntimeDependencies>
        <Dependency release="current">xorg-server-common</Dependency>
    </RuntimeDependencies>

The "current" value will be replaced by the release of the source
package that this spec file defines. This is done when building
the package so that binary packages and index files do not include
"current" string.